Moving another batch of tests.
